refactor(web): extract PostgresPlatformUserStore and Argon2id PHC validation
PostgresWebAuthService.ts was 810 lines (over the 600-line agent-check limit,
baselined at 616). Split by responsibility:
- PostgresPlatformUserStore.ts: the platform-user methods (assignPrivateDatabase,
  upsertPlatformUser, adoptPlatformUserBySecretRef) plus the hosted-status
  mapping and the platform-disabled password sentinel.
- argon2-phc-validation.ts: the Argon2id PHC-string shape/policy validators,
  re-exported from PostgresWebAuthService so the import surface is unchanged.
PostgresWebAuthService.ts is now 555 lines; its baseline entry is removed.
provision-user (CLI) constructs the store for assignPrivateDatabase; the auth
service tests use the store for the platform-user cases.

Also reconcile the public-snapshot.ts baseline 618->642 to match the
cherry-picked hosted-DB-foundation license-matrix fix (foundation growth, not
#290 code), so make agent-check exits 0.
